+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Fixed Vyos lan switch attachments; created more than needed when similar named networks existed
 - Changed resources to lowercase; easier readability in Azure
 - Fixed vyos reset function; disabled function for region 2
+- Fixed ssh keygen process; less password prompts
+- Fixed require module statement for Az; must specify individual modules.
 
 ## 1.3.2 - November 26, 2021