+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change log for AzureSite2SiteVpnLab
 
 
+
+## 1.3.4 - November 28, 2021
+
+- Added dns and dhcp option in config; autobuilds values for azure vm
+- 
+
 ## 1.3.3 - November 27, 2021
 
 - Simplified Azure connection; removed Connect-AzureEnvironment function; fixed azure subscription selection